rm rake/clean from rakefile, rm `clean` task
In this codebase, the generated lexer and parser are both commited, and
required in the gemspec for use at runtime.

We did have a `clean` task to remove all generated files - this worked
fine until the generated files were added to the gemspec, at which point
running any sort of `gem` command would error with something like

    $ gem build msh.gemspec
    WARNING:  See http://guides.rubygems.org/specification-reference/ for help
    ERROR:  While executing gem ... (Gem::InvalidSpecificationException)
        ["man/man1/msh.1"] are not files

The solution here is to force rake to rerun all dependencies

    $ bundle exec rake -h | grep B,
        -B, --build-all                  Build all prerequisites, including those which are up-to-date.

    $ bundle exec rake -B

This allows us to keep the generated files in the gemspec (since we want
the minimum amount there for runtime), while also ensuring that our
generated files are up to date.